+++ src/module/actor/sheets/compactsheet.ts
@@ -37,7 +37,7 @@ export class IronswornCompactCharacterSheet extends ActorSheet {
     return [
       {
         class: 'ironsworn-open-move-sheet',
-        label: game.i18n.localize('IRONSWORN.Moves'),
+        label: game.i18n.localize('IRONSWORN.ITEMS.TypeMove'),
         icon: 'fas fa-directions',
         onclick: (e) => this._openMoveSheet(e),
       },