+++ src/module/applications/createActorDialog.ts
@@ -18,7 +18,9 @@ export class CreateActorDialog extends FormApplication<CreateActorDialogOptions>
 
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      title: game.i18n.localize('IRONSWORN.CreateActor'),
+      title: game.i18n.format('DOCUMENT.Create', {
+        type: game.i18n.localize('DOCUMENT.Actor'),
+      }),
       template: 'systems/foundry-ironsworn/templates/actor/create.hbs',
       id: 'new-actor-dialog',
       resizable: false,