+++ src/module/vue/components/buttons/btn-rollmove.vue
@@ -3,7 +3,7 @@
     @click="rollMove"
     :tooltip="
       $t('IRONSWORN.RollMove', {
-        title: props.move?.displayName,
+        title: props.move?.moveItem().name,
       })
     "
     class="action-roll move-roll"