+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- Under the hoood: update Dataforged import with documents built in Foundry v10 ([#601](https://github.com/ben/foundry-ironsworn/pull/601))
+
 ## 1.20.13
 
 - Avoid overriding move names with Dataforged data, so translation modules can ship move names.