+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next Release
 
+- Allow pre-setting of challenge dice, and revamp the advanced rolling options UI ([#606](https://github.com/ben/foundry-ironsworn/pull/606))
+- Under the hoood: update Dataforged import with documents built in Foundry v10 ([#601](https://github.com/ben/foundry-ironsworn/pull/601))
+
 ## 1.20.13
 
 - Avoid overriding move names with Dataforged data, so translation modules can ship move names.