+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow pre-setting of challenge dice, and revamp the advanced rolling options UI ([#606](https://github.com/ben/foundry-ironsworn/pull/606))
 - Improvements to the Polish translation ([#595](https://github.com/ben/foundry-ironsworn/pull/595))
 - Under the hood: update Dataforged import with documents built in Foundry v10 ([#601](https://github.com/ben/foundry-ironsworn/pull/601))
+- Pruned, updated, and reorganized many localization keys, and borrow from FVTT's own translations to reduce the work required for future translations ([#596](https://github.com/ben/foundry-ironsworn/pull/596))
 
 ## 1.20.13